ui: extract General and User/Session tabs into ui/properties/
