+++ panel-plugin/meson.build
@@ -34,6 +34,9 @@ plugin_sources = [
   'search-page.cpp',
   'settings.cpp',
   'settings-dialog.cpp',
+  'ui/properties/common.cpp',
+  'ui/properties/general.cpp',
+  'ui/properties/user-session.cpp',
   'slot.h',
   'window.cpp',
   xfce_revision_h,
